chore: refactor money, percent, and decimal widgets into number widget
