--- app/lib/theme/widgets/ReadOnlyPercentageWidget.tsx
@@ -1,24 +1,0 @@
-import { WidgetProps } from "@rjsf/core";
-import NumberFormat from "react-number-format";
-
-const ReadOnlyPercentageWidget: React.FC<WidgetProps> = ({ id, value }) => {
-  return (
-    <dd>
-      {value ? (
-        <NumberFormat
-          fixedDecimalScale
-          id={id}
-          suffix=" %"
-          className="money"
-          decimalScale={0}
-          value={value}
-          displayType="text"
-        />
-      ) : (
-        <em>Not added</em>
-      )}
-    </dd>
-  );
-};
-
-export default ReadOnlyPercentageWidget;
